Add files via upload
Overview:
There are two files .csv file (dataset) and .ipynb (code). This project is about predicting the price of used cars. It contains a lot of analysis, machine learning models, and finally, the result which allows the user select car features, and the model will predict the price.
Title:
Used Car Price Predictor
Motivation:
Buying or selling used cars is one of the situations that you or a friend of yours have. And to protect yourself from manipulation, you need a reference that will help you find the right price for a used car in the market. This project will help two sides.
Goal:
Analyzing and understanding data
Visualize data and results
Finding the most impact features on car price.
Predict the price of the car
Data:
The dataset to be used is the used cars dataset which contains 7906 instances and 18 columns. It contains numerical and categorical variables. The dataset is from the Kaggle website: https://www.kaggle.com/datasets/shubham1kumar/usedcar-data?resource=download
Methods:
This work was done with help of Python (Panda, Numby, Matplotlib and sklearn). Describing and discovering data methods are used. Cleaning and features engineering methods are used to prepare data for processing and finding the results and goals.
Conclusion:
The most impact factors in the target (price) as a positive impact are max-power, engine, and year. km-driven factor affects the price negatively. Also from data analysis and visualization, the price increases with automatic transmission more than with manual. Price increases if the owner of the car is the first owner or if the vehicle is test-driven car. Petrol and Diesel fuel increase the price more than other fuels. 
Finally, the research question: What is the price of a car with these features (year = 2016, engine = 1248, maximum power = 88)? predicted is done correctly.